Data device.

+++ interfaces/mce_library/context.h
@@ -34,7 +34,7 @@ typedef struct mcedata {
     int (*connect)(mce_context_t);
     int (*disconnect)(mce_context_t);
     int (*ioctl)(mce_context_t, unsigned long int, int);
-    int (*read)(mce_context_t, void*, size_t);
+    ssize_t (*read)(mce_context_t, void*, size_t);
     void *write; /* laziness */
 } mcedata_t;
